+++ docusaurus/sidebars.js
@@ -1,0 +1,6 @@
+module.exports = {
+  someSidebar: {
+    Introduction: ["welcome-to-electrode", "community", "glossary"],
+    Quick_Start: ["getting-started"]
+  }
+};