+++ packages/electrode-react-webapp/lib/load-handler.js
@@ -20,9 +20,27 @@ const notFoundLoadTokenModule = m => {
   return () => ({ process: () => `\ntoken process module ${m} not found\n` });
 };
 
-module.exports = (path, tmplDir) => {
-  return optionalRequire(requireAt(Path.resolve(tmplDir || "")))(path, {
+module.exports = (path, tmplDir, customCall) => {
+  const tokenMod = optionalRequire(requireAt(Path.resolve(tmplDir || "")))(path, {
     fail: e => failLoadTokenModule(path, e),
     notFound: () => notFoundLoadTokenModule(path)
   });
+
+  if (typeof tokenMod === "function") {
+    return tokenMod;
+  }
+  if (tokenMod.tokenHandler) {
+    return tokenMod.tokenHandler;
+  }
+  if (tokenMod.default) {
+    return tokenMod.default;
+  }
+
+  if (customCall && tokenMod[customCall]) {
+    return tokenMod;
+  }
+  throw new Error(
+    `electrode-react-webapp: token module invalid - should export a function directly \
+or as 'default' or 'tokenHandler'`
+  );
 };