+++ packages/electrode-react-webapp/test/spec/hapi17.jsx-index.spec.js
@@ -3,7 +3,6 @@
 /* eslint-disable quotes */
 
 const Fs = require("fs");
-const Promise = require("bluebird");
 const assign = require("object-assign");
 const electrodeServer = require("electrode-server2");
 const Path = require("path");
@@ -13,6 +12,7 @@ const React = require("react");
 const Helmet = require("react-helmet").Helmet;
 const { runFinally, asyncVerify } = require("run-verify");
 const Munchy = require("munchy");
+const xaa = require("xaa");
 
 const getConfig = () => {
   return {
@@ -2062,7 +2062,8 @@ describe("hapi 17 electrode-react-webapp with jsx template", () => {
 
         updateTestContent(testContent1);
 
-        return Promise.try(makeRequest)
+        return xaa
+          .wrap(makeRequest)
           .then(res => {
             expect(res.statusCode).to.equal(200);
             expect(res.result).to.contain("<title>Electrode App</title>");