+++ packages/subapp-server/lib/fastify-plugin.js
@@ -22,6 +22,12 @@ module.exports = {
 
     const { routes, topOpts } = searchRoutesFromFile(srcDir, pluginOpts);
 
+    const reporting = _.get(topOpts, "reporting", {});
+    if (!reporting.enable || !reporting.reporter) {
+      // eslint-disable-next-line
+      console.warn(`Warning: Metric reporting for ssr not enabled or no reporter specified.`);
+    }
+
     const subApps = await subAppUtil.scanSubAppsFromDir(srcDir);
     const subAppsByPath = subAppUtil.getSubAppByPathMap(subApps);