+++ packages/xarc-create-app/src/check-dir.js
@@ -2,13 +2,13 @@
 const Fs = require("opfs");
 const prompts = require("prompts");
 
-async function checkDir() {
+async function checkDir(dirName) {
   const existDirFiles = await Fs.readdir(process.cwd());
   if (existDirFiles.length > 0) {
     const response = await prompts({
       type: "confirm",
       name: "overwrite",
-      message: "Your directory is not empty, write to it?"
+      message: `Your directory '${dirName}' is not empty, write to it?`
     });
 
     return response.overwrite;