+++ packages/xarc-create-app/template/_package.js
@@ -30,18 +30,18 @@ module.exports = (base, merge) => {
       npm: ">= 6"
     },
     dependencies: {
-      "@xarc/app": "^8.1.1",
+      "@xarc/app": "^8.1.6",
       "@xarc/fastify-server": "^1.1.0",
-      react: "^16.13.1",
       "react-dom": "^16.13.1",
       "react-redux": "^7.2.0",
-      redux: "^4.0.5",
-      "subapp-react": "~0.0.18",
-      "subapp-redux": "^1.0.27",
-      "subapp-server": "^1.2.1"
+      "subapp-react": "^0.0.21",
+      "subapp-redux": "^1.0.30",
+      "subapp-server": "^1.2.4",
+      react: "^16.13.1",
+      redux: "^4.0.5"
     },
     devDependencies: {
-      "@xarc/app-dev": "^8.1.1"
+      "@xarc/app-dev": "^8.1.6"
     }
   };