+++ samples/x-stylus-sample/lib/demo2/subapp-demo2.js
@@ -46,7 +46,7 @@ const Demo2 = props => {
     style: {
       textAlign: "center"
     }
-  }, "\xA9 ", new Date().getFullYear(), " Your (Company) Name Here"));
+  }, "\xA9 ", new Date().getFullYear(), " Your (Company) name here"));
 };
 
 const mapStateToProps = state => state;