+++ samples/x-stylus-sample/src/demo2/subapp-demo2.js
@@ -34,7 +34,7 @@ const Demo2 = props => {
         &nbsp;{value}&nbsp;
         <button onClick={() => dispatch(incNumber())}>&#8811;</button>
       </div>
-      <p style={{ textAlign: "center" }}>© {new Date().getFullYear()} Your (Company) Name Here</p>
+      <p style={{ textAlign: "center" }}>© {new Date().getFullYear()} Your (Company) name here</p>
     </div>
   );
 };